+++ NEWS.md
@@ -4,6 +4,8 @@
 
 - Added a global option `blogdown.server.first`, which can be specified as a function to run before serving the site. For example, you can sync a JS file to the `static/` directory with `options(blogdown.server.first = function() { file.copy('../foo/bar.js', './static/js/', overwrite = TRUE) })` each time before you serve the site.
 
+- Added arguments `arch` and `os` to `install_hugo()` so that users can choose the architecture and operating system names manually. For example, `blogdown::install_hugo(extended = FALSE, arch = 'ARM64', os = 'FreeBSD')` would install `hugo_*_FreeBSD-ARM.tar.gz` (`*` denotes a version number) from https://github.com/gohugoio/hugo/releases.
+
 ## MINOR CHANGES
 
 - The intermediate files `.knit.md$` and `.utf8.md$` no longer need to be ignored in `ignoreFiles` in the Hugo config file now. There is no harm to ignore them anyway. If you have ignored them, you do not need to update your config (#609).